move team files to team component; fix grayscale
