+++ src/components/CulturePage/index.tsx
@@ -5,7 +5,7 @@ import { atMinDesktop, atMinTablet, spacing } from '~/theme';
 
 import { Layout } from '../Layout';
 import { PageHero } from '../PageHero';
-import { TeamCardList } from '../TeamCardList';
+import { TeamCardList } from './TeamCardList';
 import { TeamHero } from '../TeamHero';
 import { AvailablePositions } from './AvailablePositions';
 import { OurBlog } from './OurBlog';
