+++ src/components/index.ts
@@ -19,9 +19,9 @@ export * from './MobileMenu';
 export * from './NetJetsPage';
 export * from './PageHero';
 export * from './QuotedImage';
-export * from './TeamCardList';
+export * from './CulturePage/TeamCardList';
 export * from './TeamHero';
-export * from './TeamMemberCard';
+export * from './CulturePage/TeamMemberCard';
 export * from './Text';
 export * from './useAnimationWhileVisible';
 export * from './useMouseAnimationWhileVisible';
